Actualizo mapas. En el Paquete 'PlayWave' cambio el tipo de paquete del wav de byte a integer. Corrijo pequeño fallo en la compresion en los mapas donde buscaba el archivo 'Maps' en lugar de 'Mapas'

+++ CODIGO/Red/Protocol.bas
@@ -2949,7 +2949,7 @@ Private Sub HandlePlayWave()
     Dim srcX As Byte
     Dim srcY As Byte
     
-    wave = incomingData.ReadByte()
+    wave = incomingData.ReadInteger()
     srcX = incomingData.ReadByte()
     srcY = incomingData.ReadByte()
